requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp==5.3.1
    # via kombu
billiard==4.2.4
    # via celery
celery==5.5.1
    # via -r requirements.in
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage[toml]==7.13.5
    # via pytest-cov
iniconfig==2.3.0
    # via pytest
kombu==5.5.4
    # via celery
packaging==26.0
    # via
    #   kombu
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
prompt-toolkit==3.0.52
    # via click-repl
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==7.1.0
    # via -r requirements.in
pytest-django==4.12.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
six==1.17.0
    # via python-dateutil
tzdata==2025.3
    # via kombu
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.6.0
    # via prompt-toolkit
